[TASK] gitlab-ci: Activate check permissions script
Since moving to gitlab, script checkFilePermissions.sh
was inactive due to a gitlab detail that leads to
funny permissions of cloned files.

This can be suppressed with a gitlab feature toggle.

The patch activates that toggle, activates the test
script and fixes a couple of files that slipped in
with broken permissions meanwhile.

Change-Id: I2b0fdca282dbc39fe230b33910a690118a21b74e
Resolves: #95194
Releases: master, 10.4, 9.5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71035
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>