[BUGFIX] Fix TypoScriptFrontendController initialization in subrequests
With the introduction of #94402 error pages are fetched via a
sub-request. Manual "page not found" ErrorController invocations
within an extbase action then resulted in rendering the originally
requested page instead of the defined 404 page.
This happened because the PrepareTypoScriptFrontendRendering
middleware hold a reference to the outer TSFE instance which
contains the original page id.

Instead of injecting the stateful TSFE (which is generally discouraged),
TSFE in now passed as request attribute. The container will
log an according warning message from now on.

Resolves: #95174
Releases: master
Change-Id: Ieda58e2bef8f08762fcba06b76df03aff7b10d5c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71084
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>