[BUGFIX] Increase z-index of element browser components
The file listing in the link browser is using bootstraps'
"List group" component. In case a list group item is
active, it receives "z-index: 2", to slightly overlay
the neighbor items.

To avoid conflicts with element browser components,
which also use "z-index: 2", this patch increases the
z-index of those components to "3".

Resolves: #95323
Releases: master
Change-Id: Ibd18992797cc8500518f589d4bebed9640b318f8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71217
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>