[TASK] Deprecate page.includeCSS/includeCSSLibs.my-identifier.import
The option to use the `@import` syntax for including
external CSS files through TypoScript has been deprecated.

This was possible through:

page = PAGE
page.includeCSSLibs.file1 = fileadmin/benni.css
page.includeCSSLibs.file1.import = 1

Through the "import = 1" option the output was

<style>
 @import url('fileadmin/benni.css');
</style>

instead of using the <link> tag.

Using the link tag, which is the de-facto syntax these days,
allows to load a file directly when interpreting the HTML of the
browser, instead of first interpreting the HTML, then the CSS
and have a blocking call to an external URL to continue interpreting
the CSS.

Resolves: #95349
Releases: master
Change-Id: I05071d800fb938af00695005081bfbfaeb4ecee5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71231
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>