[BUGFIX] Remove test data mangling in PHP 8.1 for SQLite
SQLite returns the correct type information now, so this workaround
is not necessary.

As long as we are not on php 8.1 as min requirement, use a version
check to disable the workaround for sqlite and php >8.1.

After 8.1 as minimum requirement the workaround method should be
removed at all.

Resolves: #95153
Releases: master
Change-Id: I8545d9eb847eacd13a4514034fab37cdb9a5d192
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71238
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>