[BUGFIX] Clarify requested token vs. missing token in Backend deeplinks
This change separates between a missing CSRF token and an invalid
request token in the backend redirect functionality, ensuring
that the token is given (= redirect) instead of wrong (= redirect
to log in page).

In addition, the redirect links are now only generated if a route
is not limited to POST (this was added in 11.1 later-on).

Resolves: #95370
Releases: master
Change-Id: I6ccb5dc00752d185b640de1be52b98cd4c5c10b1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71317
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>