[TASK] Clean up EXT:linkvalidator CSS and JavaScript
The EXT:linkvalidator module specific stylesheet is
removed in favor of using the bootstrap framework.
Besides increased consistency this also leads to less
custom classes and old fashioned styles.

The JavaScript component of the module is refactored
to avoid jQuery, while the code is simplified and type
hinted.

Additionally, some improvements to the module
are done. For example, proper spacing and usage
of the toggle switches.

Resolves: #95381
Releases: master
Change-Id: Ia05199a17a90087bbd9181a79ff7f1c2f16f2c4d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71338
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>