[BUGFIX] Fix issues with drag uploader and empty file lists
Previously, when a folder was empty or no files
were found for a search term, the filelist still
displayed the table header. This was related to
the Drag&Drop functionality, which internally
reads data from the table header.

However it's also sufficient to just hide the
table header and to show proper messages, including
direct actions, e.g. "reset search" or "upload files".

Therefore, also the Drag&Drop functionality is
improved. The draggable area now covers the whole
module, so the user does no longer have to strike
a 70 pixels box.

Besides, a couple of bugs in the Drag&Drop
implementation are fixed:

- When uploading files in "active search" mode, the
  file path is now properly added to the row
- After all files were uploaded, the user is now
  able to reload the module via a notification action
- Drag&Drop in FormEngine does no longer add the
  selector / checkbox field to the row
- When dragging files in and out of the dropzone,
  the dropzone is now hidden again after "dragleave"
  (only for filelist)
- The dropzone is now also hidden on "drop" (uploading
  the dragged files)
- The dropzone is not "activated" when not in the
  current visible scope (e.g. in another FormEngine tab)
- The checkbox is now also activated for uploaded
  files (without module reload)
- The full sized drozone can be closed with the ESC key
- JavaScript SyntaxError, triggered by a server error,
  such as "POST Content-Length" exceeds limit, is
  now properly handled

Resolves: #95403
Resolves: #95421
Releases: master
Change-Id: Id9de5065544aa57deaa11b8b4f9158b869a72427
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71375
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>