[TASK] Centralize Backend Redirects for Backend Routing
The internal logic to create backend links with a redirect

The internal method "UriBuilder->buildUriWithRedirectFromRequest()"
is removed, and a new RouteRedirect object is now added which
cleans up the Routes and Parameters and does various checks
in a central place.

The method "UriBuilder->buildUriWithRedirect()" now needs
a RedirectRoute object to build the proper URI.

The BackendController also checks now properly via the RouteRedirect.

Resolves: #95407
Releases: master
Change-Id: I76658b446282fe6d72b5281ed19d12f91639a355
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71380
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>