[BUGFIX] Fix the generics annotations for ObjectManager(Interface)
To avoid PHPStan warnings in projects that use ObjectManagerInterface
as a type annotation, the ObjectManager(Interface) generics need to
be on the generating methods, not on the class/interface itself.

Also streamline the PHPDoc to be more consistent, and fix the semantics
(there are no object instances, only class instances) to avoid
confusing the reader.

Releases: master, 10.4

Resolves: #95411
Change-Id: Iebc6e3c5fbe763de5d9329c0f4351bfb5ba0dd1a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71382
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>