[BUGFIX] Fix the generics annotations for ObjectManager(Interface)
To avoid PHPStan warnings in projects that use ObjectManagerInterface
as a type annotation, the ObjectManager(Interface) generics need to
be on the generating methods, not on the class/interface itself.

Also streamline the PHPDoc to be more consistent, and fix the semantics
(there are no object instances, only class instances) to avoid
confusing the reader.

Resolves: #95411
Releases: master, 10.4
Change-Id: Iebc6e3c5fbe763de5d9329c0f4351bfb5ba0dd1a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71465
Tested-by: core-ci <typo3@b13.com>
Tested-by: Elias Häußler <e.haeussler@familie-redlich.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Elias Häußler <e.haeussler@familie-redlich.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>