[BUGFIX] Fix extbase Request->withoutAttribute('extbase')
Avoid wrapping the request with the extbase request wrapper which
enforces the availability of the extbase attribute and thus causes
an InvalidArgumentException if the attribute has been removed.

Now that the usecase to remove the extbase attribute is fixed
via the PSR-7 standard method ->withoutAttribute('extbase'),
the misleading method Request::getServerRequest() is removed as it
is supposed to return the original request which it actually didn't do.

This method implies that the server ("original"?) PSR-7 request will be
returned, but that isn't true. Instead a cloned object with the
extbase attribute but without the extbase wrapper is returned.
That's counter intuitive and therefore the method is pretty
useless and removed.

The usecase for getServerRequest() is marginal, the Request
is a ServerRequestInterface anyway, therefore it can be
used and passed everywhere where ServerRequestInterface is
required and would therefore only be useful to remove the
extbase attribute (which was broken prior to this patch).

Releases: master
Resolves: #95422
Related: #94428
Change-Id: Ia543a70ecc990b044a37570a91fff427627df941
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71396
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>