[BUGFIX] Restore Drag&Drop functionality in page module
The Drag&Drop component previously misused the
new content element wizard related “t3js” class to
check whether content can be created or not. This
had two side effects:

* Drag&Drop was never possible when new content
  element wizard was disabled
* Since #95277 Drag&Drop was completely unusable,
  because the mentioned class got removed in favor
  of a new web component

We already had the same issue with the paste button
in #95117. Therefore this is now fixed in the same way.

In the fluid based page module, the  `.t3js-page-new-ce`
class is always present, in case content creation is allowed.
Therefore it's sufficient to just check for this class. This
also makes the `else` part superfluous.

Conclusion: Drag&Drop is now always working, no matter
if new content element wizard is enabled or not, as long
as content creation is allowed for this column (e.g. no
access restricted colPos).

Resolves: #95440
Releases: master
Change-Id: I4d07d523f3b8394edccc88dbc414876f9b33ab53
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71411
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Helmut Hummel <typo3@helhum.io>