[BUGFIX] Ensure subfolder backend routing work for Apache
With #93048 url rewrites for backend links have been introduced,
which required additional rewrite rules in the root htaccess.

The used rewrite rules for Apache route to the frontend end
index.php for installations in subfolders, thus breaking
backend usage.

This patch change the needed rewrite rule to route correctly
to the backend index.php for backend links. Correct rule is
placed in the root-htaccess template and the webserver
configuration write service is adjusted to address this for
old configs too. Furthermore the breaking rst file is adjusted
to point out the correct needed rewrite rule.

The RewriteCond statement is removed as the condition
is fully covered by the RewriteRule itself.

Resolves: #94054
Related: #95180
Related: #93048
Releases: master
Change-Id: I52870f5e5a805c2b00e8ad146eb8ecd009811a68
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71422
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>