[BUGFIX] Allow FilePathSanitizer to return absolute paths
There are few cases where the resulting path from FilePathSanitizer
is needed and used as absolute path to a private file.

For these usages the file now is returned early.

Additionally the logic to make a relative path out of
an absolute path is now relaxed a bit, so that legacy
usages of this API won't fail any more.

An exception is still thrown, when a path from Composer
package is requested, that can not be resolved to a public path.

Resolves: #95481
Resolves: #95495
Releases: master
Change-Id: I8df7d74e6e71154be0508a96cdc19821beeb5ecb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71457
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>