[TASK] Skip one unit test dataset only for php81 until php bug is fixed
PHP8.1.0 changed internally from gethostbyaddr() to gethostinfo() from
standard c library for the php gethostbyaddr().

On alpine docker images this leads to an issues with ::1 hostname
resolving in a unit test.

This is out of the scope to be handled by the core, thus this patch
disabled this single dataset only for 8.1, adding a todo for recheck
when the php bug is solved or otherwise mitigated.

Corresponding Issue Tracker Links:

https://github.com/docker-library/php/issues/1201
https://bugs.php.net/bug.php?id=81441

Resolves: #95501
Releases: master
Change-Id: I97b60dae3cb0186124c655a529518fa93b3d8b7c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71470
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>