[BUGFIX] Restore inconsistent language highlighting
The page module displays a flash message for page
translations with inconsistent ("mixed mode") content
elements. The last sentence of the flash message reads:
"We highlighted the problematic records for you". This
however was no longer the case in the fluid based page
module.

Therefore, this patch restores highlighting of such records.

Resolves: #91177
Releases: master
Change-Id: Ic0ada77428a426254f9cbba256e23f938733f9aa
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71484
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Rachel Foucard <rfoucard@w-seils.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Rachel Foucard <rfoucard@w-seils.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>