[BUGFIX] Properly store and retrieve task group collapse state
In #86291 the scheduler task list was refactored. A
part of this refactoring was aligning the list with the
one, known from the recordlist module. Therefore
also the "toggle table" functionality from EXT:recordlist
was added.

However, since the updated state was never retrieved
on building the table, the collapse state was saved
but never used.

While integrating the multi record selection in #94918,
the corresponding "t3js-toggle-recordlist" class was
removed to prevent cross dependencies. Since then,
the collapse state was not just not retrieved, but no
longer saved at all.

This is now fixed by properly implementing this
functionality in the corresponding scheduler
JavaScript module, while the saved state is
now properly taken into account on building
the table.

Resolves: #88351
Related: #86291
Related: #94918
Releases: master
Change-Id: Ie8cb870bfba8170adec98410b6de2e53bbcf7938
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71506
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>