[BUGFIX] Properly set softRefParams in LinkAnalyzer
This correctly sets "subst" as a value of the
softRefParams array. Previously "subst" was
set as key, which only worked because of a
non strict "in_array" check, which also
returned TRUE for this case, but then broke
with the refactoring in #94687.

Additionally the regular expression of the
UrlSoftReferenceParser, used to find urls
in content, also got extended in #94687.
This is now partially reset to not longer
match links in tag attributes. This otherwise
would lead to duplicate entries, since those
urls are already matched by the
TypolinkTagSoftReferenceParser.

The impact of the changed regular expression
was overlooked beforehand, due to the bug
in LinkAnalyzer.

Resolves: #95449
Related: #94687
Releases: master
Change-Id: I3ca134cd1eed95b8563335708e393f5b5fa79420
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71511
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>