[BUGFIX] Align default value of displayThumbs
The "displayThumbs" module setting is
fetched two times in the FileBrowser. One
time to determine whether thumbnails
should be displayed and a second time
to set the current value of the display
thumbnails "toggle".

However, the default values of those two
cases differed, which then led to a default
state where the "toggle" was active, while
no thumbnails were displayed.

This is now fixed by using the same default
for both cases.

Resolves: #95425
Releases: master
Change-Id: I9cdf81a28aceb91d5f419332d24337b22de2a575
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71515
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>