[BUGFIX] Move virtual storage 0 back to public path
It was a conceptual mistake to move storage 0 base path
outside the document root (public path), because this complete
compatibility layer relies on the fact that this storage
is public. Moving it outside the public path, implicitly
made this storage private, with all logical consequences like
eID public URLs. This of course isn't desired and a breaking
change of behaviour.

With moving the base path back it means, that trying to access this
BC layer with paths that are not within the public path,
will not work any more. This is fine though, because this
never worked before TYPO3 11.5 anyway.

The reason why this was made working was EXT:form, which
used the storage 0 compat layer for resolving form
definitions from extensions, which did not work, when
installing extensions into vendor folder (instead of the still
current default, where extensions are installed in public folder).

Accessing public resources from vendor installed extensions
will still work with this change, because the public asset
URL will automatically be calculated, but accessing private resources
will thrown an exception to clarify that the storage 0 compat layer
is not made to make this possible.

Also fix form framework to not rely on FAL compatibility
layer during form loading.

Resolves: #95543
Relates: #95437
Releases: master
Change-Id: Ie218dbb8ace9999f2bead77b671ae87f6ad46170
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71512
Tested-by: core-ci <typo3@b13.com>
Tested-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>