[TASK] Add class name to deprecated cObj for plugins message
Public property $cObj for plugins has been deprecated in
v11 with #94956 and setContentObjectRenderer() method has
been introduced as replacement.

If this method is missing, a deprecation message is triggered.

This patch inserts the class name of the plugin into the message,
to make it easier to determine where the new method is missing.

Resolves: #95554
Releases: master
Change-Id: Iffae1fcbd2f59bd4fe353e585dd5f41914bcc760
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71524
Reviewed-by: Stephan Großberndt <stephan.grossberndt@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>