[BUGFIX] Check for TSFE before use in ContentObjectRenderer
TypoScriptFrontendController has not been injected to a
ContentObjectRenderer instance in a test, thus failing
with disabled error handler. TypoScriptFrontendController
dependency is based on needed access to initialized
TypoScript configurations.

This patch checks if tsfe is available before accessing
properties and methods in ContentObjectRenderer to
mitigate this warning instead of mocking an uninitialized
TSFE instance in the test.

Resolves: #95610
Releases: master
Change-Id: Id3e8c2b7029ee0932da242c47ffb1a3837897bd8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71600
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>