[BUGFIX] Restore "edit metadata" of translated files in FormEngine
In case a files' metadata were translated in a L>0 language,
whenever the file is references in a record with this L>0
language, the inline controls show the "edit metadata"
button again.

This previously did not work, because the language field,
resolved by FormEngine, might not always be an array with
the language uid on [0]. Other FormEngine parts already
implement proper checks. So does the InlineRecordContainer
now.

Additionally, this also fixes an undefined array key access
warning and replaces the hardcoded "sys_language_uid"
field name, since this might be changed in TCA.

Resolves: #95617
Releases: master
Change-Id: I98f0ec656a2296b6a78c15c8c6aecd0667745ae0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71608
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>