[BUGFIX] Prevent exceptions for new user accounts
In case new user accounts are not properly set up
and the user tries to login, some toolbar related
components led to exceptions.

This is now fixed by taking the current user status
into account. In detail, the workspace title is only
checked, if the user has a proper workspace defined
and the user module children are only checked if
the user has access to the main module.

Resolves: #95336
Releases: master
Change-Id: I64b05b5b04298ea17756f20ddd2fe97de558c825
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71627
Tested-by: Jonas Eberle <flightvision@googlemail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>