[BUGFIX] Avoid warning in ResourceStorage
ResourceStorage->getUniqueName() is used to return a
unique and sanitized file or folder name. It uses
pathinfo() on the given name. pathinfo() returns the
'file ending' (like 'txt') as 'extension' array key. This
array key is not set when a file without 'extension' is
given (like "foo"). This leads to a warning getUniqueName().

The patch sanitizes access to 'extension' using null
coalescing operator.

Resolves: #95637
Releases: master
Change-Id: I3a834c910c0fbf7d2c8640fc917284c9b229face
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71635
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>