[BUGFIX] Consider table name on retrieving extension key
The NewRecordController previously relied on a record
title being a locallang reference. This was used to
determine the extension key, to further fetch the
corresponding extension title and the icon.

This then lead to various misbehaviour for records,
not using a locallang reference as title. For example,
the last record title might end up being used as the
group title, and similar.

To prevent such scenarios an additional check is
added, which tries to fetch the extension key from
the table name, which might most probably work
since by convention a table name starting with
tx_ or tt_ is followed by the extension name.

Another topic, not part of this patch is the fact
that the extension title is fetched from the
ext_emconf.php file. This file however might
no longer exist in the future. Therefore, the
corresponding part needs to be refactored
to also take the composer.json into account.

Resolves: #94049
Resolves: #95641
Releases: master
Change-Id: Ic5e27c68571fc2a1a6fe19c0017b26795dbebec0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71640
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>