[BUGFIX] Avoid dangling error handler registrations in tests
The functional ErrorHandlerTest uses set_error_handler()
to register test subjects. These need to be dropped
afterwards to not end up with additional error handlers
in other tests which may suppress errors we'd see otherwise.

The testing-framework will soon have a check to avoid
this scenario in the future, so this bug can't happen
again.

Change-Id: I6861932807c9f90efb5a7c1e233e4bafd9674ab1
Resolves: #95649
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71752
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>