[BUGFIX] Avoid incomplete setup in StorageRepositoryTest
The StorageRepositoryTest tests createLocalStorage() which
calls testCaseSensitivity() to find out if a local filesystem
is case sensitive. It creates a test file to do so, but does
not check if the directory itself exists.

This should probably handled in createLocalStorage(), but
for now we create the directories in the test to fix
a raised warning and add a @todo to test and code.

Resolves: #95651
Releases: master
Change-Id: I38455b7233c13607b4756d5c356e83915b8d012d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71772
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>