[BUGFIX] Do not log page access failures
It does not make sense to log page access failures
to the logging framework. Pages are inaccessible
for a reason, no need to log every try.
(404 responses are logged by the webserver)

This helps with monitoring and removes useless
log messages from the logs.

Also avoid to respond with an internal server error to
page-requests that are hidden by (inherited) rootline
constraints. These are normalized to 404. These cases
should be improved to differentiate between 404 (hidden
due to rootline) and 403 (access denied due to rootline)
in future.

Releases: master, 10.4
Resolves: #93851
Change-Id: Id9c1c886e042d20276fbe241f26505765d924fec
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71797
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Christoph Lehmann <christoph.lehmann@networkteam.com>
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>