[BUGFIX] Avoid deprecated methods in DebuggerUtility
Extbase DebuggerUtility uses reflection to determine
information that should be rendered as debug output
in 'renderContent()'.

PHP 8.0 deprecated some methods of ReflectionParameter
which will be removed in PHP 9.0.

This patch avoids the deprecation message using the
alternative way to get these informations.

Additionally tests are added to cover the behaviour.

Resolves: #95704
Releases: master
Change-Id: I808f88b912d41b50ad551bb3411818f797badbbc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71892
Tested-by: core-ci <typo3@b13.com>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>