[BUGFIX] Split ContainerTest fixtures per class
Having more than one class per file is discouraged
and leads to issues with class-alias-loader
if non-namespaced classes are handled.

The test classes are target of test subject
extbase Container, which is deprecated. We don't
care that too much and simply split file
Testclasses.php into single files to get rid of
the issue.

Resolves: #95707
Releases: master
Change-Id: I61edf5abf934d3d0a92e62bd97cc3a16415d5590
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71895
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>