[BUGFIX] Fix translation handling in vendor packages
When installing TYPO3 and Extensions into the vendor folder, no language
files can be loaded anymore because of an implementation depending on
hard coded Extension directories.

A new method is implemented in the PackageManager to extract a package
key from a path which solves the issue and makes the old implementation
superfluous. The code is left there however to not interfere with custom
implementations (despite these classes are marked internal).

LLL references not using an EXT: path, but a file system path will
"work" exactly like before. This should rather be deprecated later on.

In fact the whole language file parsing deserves a complete overhaul, which
should be tackled in subsequent changes for a new major TYPO3 version.

Resolves: #95839
Releases: master
Change-Id: I5ca698d445ae6c5125ee42787fd7274dc8d27a38
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71938
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Benni Mack <benni@typo3.org>