[BUGFIX] Avoid passing null parameter in SchemaColumnDefinitionListener
SchemaColumnDefinitionListener instantiates a new Column class and
passes the column name retrieved from an array. If the array key
'field' does not exist in the tableColumn definition, the null
coalescing operator with the fallback of 'null' is used, which is not
ideal because the fieldname parameter has a docblock type hint of
string.

DBAL internally there is a check of the column name using the
php-internal function strpos(), which since PHP 8.1 causes a
deprecation message if a null value is passed as first parameter.

This patch changes the value of the null coalescence operator to an
empty string to avoid this deprecation message, and to be compatible
when DBAL changes the parameter type hint to string.

Resolves: #95743
Releases: master
Change-Id: I95869666cf1c8f798259652cf63bdd8fbfc554fd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71946
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>