[BUGFIX] Avoid deprecation message in MetaDataAspect
MetaDataAspect implements multiple interfaces and has return type hints
that match the interfaces. Except for the offsetGet() and current()
methods, which are defined by the ArrayAccess and Iterator interfaces.

The 'mixed' return type cannot be used before PHP 8.0, so it cannot be
added yet.

This patch adds the attribute '#[\ReturnTypeWillChange]' to avoid the
fatal error for PHP 8.1 and @todo comments to change the correct return
types, as a breaking patch for TYPO3 v12.

Resolves: #95744
Releases: master
Change-Id: Ic2c38758bbeaeef96efdcdf5f3349e3e3e08a72c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71947
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>