[BUGFIX] Update composer dev dependencies to min required versions
phpunit released new versions for 8.5.x and 9.5.x, which
no longer automatically throw exception for user deprecations.
Furthermore there was a patch which needs to have
'phpspec/prophecy' on ^1.14.0 as minimum requirement.

This patch increase the min version of the two dev dependencies.
The patch is not strictly needed in 10.4 and was mitigated with
other changes in testing-framework already, but it doesn't hurt
to have this in 10.4 to better align with v11.

Used composer commands:

composer req --dev --with-dependencies phpunit/phpunit:^8.5.21
composer req --dev phpspec/prophecy:^1.14.0

Resolves: #95360
Related: #95210
Releases: master, 10.4
Change-Id: Ia009fe20a0b53b3d1bdef097b00c1993d3b8c596
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71956
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>