[BUGFIX] Avoid calling trim(null) in PageRenderer
PageRender sanitizes array values in two methods, using
array walk with trim(). These values can be other
types then string, which trigger deprecation warning in
PHP 8.1.

This patch change the array walk to an closure to have
an explicit type cast to string for passing the value
to trim().

Additional a return type ': string' is added to the
protected method getTemplate(), as this method always
returns a string.

Resolves: #95788
Releases: master
Change-Id: Id0cdef928f977fdc1a9d476475bd3a44c5783922
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71984
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>